Default return values on functions when no webtag present.

+++ forum/include/links.inc.php
@@ -21,7 +21,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307
 USA
 ======================================================================*/
 
-/* $Id: links.inc.php,v 1.30 2004-04-04 21:03:40 decoyduck Exp $ */
+/* $Id: links.inc.php,v 1.31 2004-04-05 20:54:47 decoyduck Exp $ */
 
 function links_get_in_folder($fid, $invisible = false, $sort_by = "TITLE", $sort_dir = "ASC") // setting $invisible to true gets links that are marked as not visible too
 {
@@ -34,7 +34,7 @@ function links_get_in_folder($fid, $invisible = false, $sort_by = "TITLE", $sort
     if (!in_array($sort_by, $sort_array)) $sort_by = 'TITLE';
     if ((trim($sort_dir) != 'DESC') && (trim($sort_dir) != 'ASC')) $sort_dir = 'DESC';
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return $links;
 
     $sql  = "SELECT LINKS.LID, LINKS.UID, USER.LOGON, USER.NICKNAME, LINKS.URI, LINKS.TITLE, ";
     $sql .= "LINKS.DESCRIPTION, LINKS.VISIBLE, UNIX_TIMESTAMP(LINKS.CREATED) AS CREATED, LINKS.CLICKS, ";
@@ -63,7 +63,7 @@ function links_folders_get($invisible = false)
 {
     $db_links_folders_get = db_connect();
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return false;
     
     $sql  = "SELECT FID, PARENT_FID, NAME, VISIBLE FROM {$table_data['PREFIX']}LINKS_FOLDERS ";
     if (!$invisible) $sql .= "WHERE VISIBLE = 'Y' ";
@@ -89,7 +89,7 @@ function links_add($uri, $title, $description, $fid, $uid, $visible = true)
 
     $visible = $visible ? "Y" : "N";
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return false;
 
     $sql = "INSERT INTO {$table_data['PREFIX']}LINKS (URI, TITLE, DESCRIPTION, FID, UID, VISIBLE, CREATED) ";
     $sql.= "VALUES ('$uri', '$title', '$description', '$fid', '$uid', '$visible', NOW())";
@@ -105,7 +105,7 @@ function links_add_folder($fid, $name, $visible = false)
 
     $visible = $visible ? "Y" : "N";
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return false;
 
     $sql = "INSERT INTO {$table_data['PREFIX']}LINKS_FOLDERS (FID, PARENT_FID, NAME, VISIBLE) ";
     $sql.= "VALUES (NULL, $fid, '$name', '$visible')";
@@ -169,7 +169,7 @@ function links_change_visibility($lid, $visible = true)
 
     $db_links_change_visibility = db_connect();
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return false;
 
     $sql = "UPDATE {$table_data['PREFIX']}LINKS SET VISIBLE = '$visible' WHERE LID = '$lid'";
     return db_query($sql, $db_links_change_visibility);
@@ -181,7 +181,7 @@ function links_click($lid)
 
     $db_links_click = db_connect();
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return false;
 
     $sql = "UPDATE {$table_data['PREFIX']}LINKS SET CLICKS = CLICKS + 1 WHERE LID = '$lid'";
     $result_id = db_query($sql, $db_links_click);
@@ -199,7 +199,7 @@ function links_get_single($lid)
 
     $db_links_get_single = db_connect();
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return false;
 
     $sql  = "SELECT LINKS.FID, LINKS.UID, LINKS.URI, LINKS.TITLE, LINKS.DESCRIPTION, UNIX_TIMESTAMP(LINKS.CREATED) AS CREATED, ";
     $sql .= "LINKS.VISIBLE, LINKS.CLICKS, USER.LOGON, USER.NICKNAME, AVG(LINKS_VOTE.RATING) AS RATING, COUNT(LINKS_VOTE.RATING) AS VOTES ";
@@ -233,7 +233,7 @@ function links_get_all($invisible = false, $sort_by = "DATE", $sort_dir = "DESC"
 
     $db_links_get_in_folder = db_connect();
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return $links;
 
     $sql  = "SELECT LINKS.LID, LINKS.UID, USER.LOGON, USER.NICKNAME, LINKS.URI, LINKS.TITLE, ";
     $sql .= "LINKS.DESCRIPTION, LINKS.VISIBLE, UNIX_TIMESTAMP(LINKS.CREATED) AS CREATED, LINKS.CLICKS, ";
@@ -265,7 +265,7 @@ function links_folder_change_visibility($fid, $visible = true)
 
     $visible = $visible ? "Y" : "N";
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return false;
 
     $sql = "UPDATE {$table_data['PREFIX']}LINKS_FOLDERS SET VISIBLE = '$visible' WHERE FID = $fid";
     return db_query($sql, $db_links_folder_change_visibility);
@@ -279,7 +279,7 @@ function links_folder_delete($fid)
 
     $db_links_folder_delete = db_connect();
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return false;
 
     $sql = "SELECT MIN(FID) AS FID FROM {$table_data['PREFIX']}LINKS";
     $result_id = db_query($sql, $db_links_folder_delete);
@@ -303,7 +303,7 @@ function links_get_vote($lid, $uid)
     if (!is_numeric($lid)) return false;
     if (!is_numeric($uid)) return false;
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return false;
 
     $sql = "SELECT RATING FROM {$table_data['PREFIX']}LINKS_VOTE WHERE LID = $lid AND UID = $uid";
     $result_id = db_query($sql, $db_links_get_vote);
@@ -324,7 +324,7 @@ function links_vote($lid, $vote, $uid)
     if (!is_numeric($vote)) return false;
     if (!is_numeric($uid))  return false;
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return false;
 
     $sql = "UPDATE {$table_data['PREFIX']}LINKS_VOTE SET RATING = '$vote', TSTAMP = NOW() ";
     $sql.= "WHERE UID = '$uid' AND LID = '$lid'";
@@ -349,7 +349,7 @@ function links_add_comment($lid, $uid, $comment)
     if (!is_numeric($lid))  return false;
     if (!is_numeric($uid))  return false;
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return false;
 
     $sql = "INSERT INTO {$table_data['PREFIX']}LINKS_COMMENT (LID, UID, COMMENT, CREATED) ";
     $sql.= "VALUES ('$lid', '$uid', '$comment', NOW())";
@@ -363,7 +363,7 @@ function links_get_comments($lid)
 
     if (!is_numeric($lid))  return false;
     
-    $table_data = get_table_prefix();
+    if (!$table_data = get_table_prefix()) return false;
 
     $sql  = "SELECT USER.UID, USER.LOGON, USER.NICKNAME, UNIX_TIMESTAMP(LINKS_COMMENT.CREATED) AS CREATED, ";
     $sql .= "LINKS_COMMENT.CID, LINKS_COMMENT.COMMENT ";
@@ -401,7 +401,7 @@ function links_delete_comment($cid)
 {
     $db_links_delete_comment = db_connect();
     
-    $table_data = get_table_prefix();
+    if ($table_data = get_table_prefix()) return false;
     
     $sql = "DELETE FROM {$table_data['PREFIX']}LINKS_COMMENT WHERE CID = $cid";
     $result_id = db_query($sql, $db_links_delete_comment);
@@ -414,7 +414,7 @@ function links_delete($lid)
 
     if (!is_numeric($lid))  return false;
     
-    $table_data = get_table_prefix();
+    if ($table_data = get_table_prefix()) return false;
 
     $sql = "DELETE FROM {$table_data['PREFIX']}LINKS WHERE LID = '$lid'";
     $result_id = db_query($sql, $db_links_delete);
@@ -433,7 +433,7 @@ function links_update($lid, $fid, $title, $uri, $description)
     if (!is_numeric($lid))  return false;
     if (!is_numeric($fid))  return false;
     
-    $table_data = get_table_prefix();
+    if ($table_data = get_table_prefix()) return false;
 
     $sql = "UPDATE {$table_data['PREFIX']}LINKS SET LID = '$lid', FID = '$fid', ";
     $sql.= "TITLE = '$title', URI = '$uri', DESCRIPTION = '$description' WHERE LID = '$lid'";
@@ -447,7 +447,7 @@ function links_get_creator_uid($lid)
 
     if (!is_numeric($lid))  return false;
     
-    $table_data = get_table_prefix();
+    if ($table_data = get_table_prefix()) return false;
 
     $sql = "SELECT UID FROM {$table_data['PREFIX']}LINKS WHERE LID = '$lid'";
     $result_id = db_query($sql, $db_links_get_creator_uid);
@@ -461,7 +461,7 @@ function links_get_comment_uid($cid)
 
     if (!is_numeric($cid))  return false;
     
-    $table_data = get_table_prefix();
+    if ($table_data = get_table_prefix()) return false;
 
     $sql = "SELECT UID FROM {$table_data['PREFIX']}LINKS_COMMENT WHERE CID = '$cid'";
     $result_id = db_query($sql, $db_links_get_comment_uid);
